requirements.txt
# requirements.txt
# Основные зависимости Crypto Scalper Bot (февраль 2026)

# Core
python>=3.10,<3.13
pyyaml>=6.0.1               # Работа с конфигами
polars>=0.20.0              # Основная обработка данных (быстрее pandas)
numpy>=1.26.0               # Массивы и математика

# Binance API и торговля
ccxt>=4.3.0                 # Библиотека для Binance (async support)
httpx>=0.27.0               # Асинхронные HTTP-запросы (внутренне в ccxt)

# Модель и обучение
torch>=2.2.0                # PyTorch (GPU/CPU/MPS)
torchvision>=0.17.0         # Не обязательно, но полезно для ONNX
onnx>=1.16.0                # Экспорт модели в ONNX
onnxruntime>=1.17.0         # ONNX Runtime (особенно mobile для телефона)

# Кластеризация и статистика
hdbscan>=0.8.33             # HDBSCAN для сценариев
scikit-learn>=1.4.0         # Для дополнительных метрик (если нужно)

# Логирование и прогресс
tqdm>=4.66.0                # Прогресс-бары
rich>=13.7.0                # Красивые логи (опционально)

# Параллелизация и производительность
joblib>=1.4.0               # Параллельный бэктест на сервере
psutil>=5.9.8               # Мониторинг CPU, памяти, температуры (телефон)

# Дополнительно (для Google Sheets, тестов, графиков)
gspread>=6.1.0              # Экспорт сценариев в Google Sheets (опционально)
oauth2client>=4.1.3         # Credentials для gspread
matplotlib>=3.8.0           # Графики (для анализа в Jupyter)
pytest>=8.0.0               # Unit-тесты (в папке tests/)

# Установка:
# pip install -r requirements.txt
# Для GPU (сервер/Colab): pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# Для телефона (Android): pip install --no-deps torch==2.2.0+cpu torchvision==0.17.0+cpu -f https://download.pytorch.org/whl/torch_stable.html